core: Split of non-head-node exclusion code
