+++ src/core/bonded_interactions/bonded_interaction_data.cpp
@@ -19,7 +19,9 @@
 #include "bonded_interaction_data.hpp"
 #include "communication.hpp"
 
+#include <boost/range/algorithm/copy.hpp>
 #include <boost/range/numeric.hpp>
+
 #include <utils/constants.hpp>
 
 std::vector<Bonded_ia_parameters> bonded_ia_params;
@@ -145,4 +147,49 @@ void remove_all_bonds_to(Particle &p, int id) {
       i += 1 + partners;
   }
   assert(i == bl->n);
+}
+
+void add_bond(Particle &p, Utils::Span<const int> bond) {
+  boost::copy(bond, std::back_inserter(p.bl));
+}
+
+int delete_bond(Particle *part, const int *bond) {
+  IntList *bl = &part->bl;
+  int i, j, type, partners;
+
+  // Empty bond means: delete all bonds
+  if (!bond) {
+    bl->clear();
+
+    return ES_OK;
+  }
+
+  // Go over the bond list to find the bond to delete
+  for (i = 0; i < bl->n;) {
+    type = bl->e[i];
+    partners = bonded_ia_params[type].num;
+
+    // If the bond type does not match the one, we want to delete, skip
+    if (type != bond[0])
+      i += 1 + partners;
+    else {
+      // Go over the bond partners
+      for (j = 1; j <= partners; j++) {
+        // Leave the loop early, if the bond to delete and the bond with in the
+        // particle don't match
+        if (bond[j] != bl->e[i + j])
+          break;
+      }
+      // If we did not exit from the loop early, all parameters matched
+      // and we go on with deleting
+      if (j > partners) {
+        // New length of bond list
+        bl->erase(bl->begin() + i, bl->begin() + i + 1 + partners);
+
+        return ES_OK;
+      }
+      i += 1 + partners;
+    }
+  }
+  return ES_ERROR;
 }
