+++ src/core/electrostatics_magnetostatics/magnetic_non_p3m_methods.cpp
@@ -28,8 +28,6 @@
 #include "electrostatics_magnetostatics/dipole.hpp"
 #include "errorhandling.hpp"
 #include "grid.hpp"
-#include "nonbonded_interactions/nonbonded_interaction_data.hpp"
-#include "particle_data.hpp"
 
 #include <utils/constants.hpp>
