+++ src/core/energy_inline.hpp
@@ -61,7 +61,7 @@
 #include "bonded_interactions/bonded_coulomb_sr.hpp"
 #include "electrostatics_magnetostatics/coulomb_inline.hpp"
 #endif
-#include "particle_data.hpp"
+#include "exclusions.hpp"
 #include "particle_index.hpp"
 #include "statistics.hpp"
