+++ src/core/forces_inline.hpp
@@ -36,6 +36,7 @@
 #include "bonded_interactions/thermalized_bond.hpp"
 #include "bonded_interactions/umbrella.hpp"
 #include "errorhandling.hpp"
+#include "exclusions.hpp"
 #include "forces.hpp"
 #include "immersed_boundary/ibm_tribend.hpp"
 #include "immersed_boundary/ibm_triel.hpp"
@@ -59,7 +60,6 @@
 #include "npt.hpp"
 #include "object-in-fluid/oif_global_forces.hpp"
 #include "object-in-fluid/oif_local_forces.hpp"
-#include "particle_data.hpp"
 #include "particle_index.hpp"
 #include "rotation.hpp"
 #include "thermostat.hpp"
