+++ src/core/grid_based_algorithms/lbgpu.cpp
@@ -36,7 +36,6 @@
 #include "grid_based_algorithms/lbgpu.hpp"
 #include "integrate.hpp"
 #include "nonbonded_interactions/nonbonded_interaction_data.hpp"
-#include "particle_data.hpp"
 #include "statistics.hpp"
 
 #include <utils/constants.hpp>
